up1
Separate the home and the Admin dashboard components
