+++ pfe-frontend/src/index.css
@@ -4,7 +4,8 @@
 
 @font-face {
   font-family: "NovaFlat-Regular";
-  src: url("/src/assets/fonts/Nova-flat/NovaFlat-Regular.ttf") format("truetype"); /* Adjust format as needed */
+  src: url("/src/assets/fonts/Nova-flat/NovaFlat-Regular.ttf")
+    format("truetype"); /* Adjust format as needed */
   font-weight: normal;
   font-style: normal;
 }
@@ -16,10 +17,12 @@
 }
 @font-face {
   font-family: "Host-grotest";
-  src: url("/src/assets/fonts/Host-grotesk/static/HostGrotesk-Regular.ttf") format("truetype"); /* Adjust format as needed */
+  src: url("/src/assets/fonts/Host-grotesk/static/HostGrotesk-Regular.ttf")
+    format("truetype"); /* Adjust format as needed */
   font-weight: normal;
   font-style: normal;
-}@font-face {
+}
+@font-face {
   font-family: "Roboto";
   src: url("/src/assets/fonts/Roboto/Roboto-Regular.ttf") format("truetype"); /* Adjust format as needed */
   font-weight: normal;
